# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tenders API Documentation",
        "description": "Tenders API Documentation",
        "contact": {
            "email": "info@just.sd"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Tenders API Server"
        }
    ],
    "paths": {
...
        "/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get list of projects",
                "description": "get all projects",
                "operationId": "get_all_projects",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "store project",
                "description": "store project",
                "operationId": "store_project",
                "parameters": [
                    {
                        "name": "project_code",
                        "in": "path",
                        "description": "Project Code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "project_title",
                        "in": "path",
                        "description": "Project Title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "path",
                        "description": "Start Date Ex. (2024-03-20)",
                        "required": true,
                        "schema": {
                            "type": "Date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "path",
                        "description": "End Date Ex. (2024-03-20)",
                        "required": true,
                        "schema": {
                            "type": "Date"
                        }
                    },
                    {
                        "name": "country",
                        "in": "path",
                        "description": "Country",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "path",
                        "description": "State",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "path",
                        "description": "City",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "path",
                        "description": "Address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "show project",
                "description": "show project",
                "operationId": "show_project",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "update project",
                "description": "update project",
                "operationId": "update_project",
                "parameters": [
                    {
                        "name": "project_code",
                        "in": "path",
                        "description": "Project Code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "project_title",
                        "in": "path",
                        "description": "Project Title",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "path",
                        "description": "Start Date Ex. (2024-03-20)",
                        "required": true,
                        "schema": {
                            "type": "Date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "path",
                        "description": "End Date Ex. (2024-03-20)",
                        "required": true,
                        "schema": {
                            "type": "Date"
                        }
                    },
                    {
                        "name": "country",
                        "in": "path",
                        "description": "Country",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "path",
                        "description": "State",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "path",
                        "description": "City",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "path",
                        "description": "Address",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "delete project",
                "description": "delete project",
                "operationId": "delete_project",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}